requirements.txt
streamlit
google-genai
py3Dmol
streamlit-screen-stats
streamlit-local-storage
streamlit-browser-session-storage
pubchempy
toml
